% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crug_skeleton.R
\name{crug_skeleton}
\alias{crug_skeleton}
\title{Project Template demo}
\usage{
crug_skeleton(path, ...)
}
\arguments{
\item{path}{The path of the new project}

\item{...}{additional project template options}
}
\description{
Project Template demo
}
